pkg/logical/message.go
Extract logical message types into their own file, and encourage people
not to write methods by writing in the style of type-only definitions.

Message types should, ideally, be plain-old-objects.
